+++ emulator/src/core/device.hpp
@@ -13,8 +13,8 @@ public:
     virtual void initialize();
     virtual void reset();
     virtual void update();
-    virtual uint64_t read(uint64_t address) const;
-    virtual void write(uint64_t address, uint64_t data);
+    virtual uint8_t read(uint64_t address) const;
+    virtual void write(uint64_t address, uint8_t data);
 
     virtual uint64_t getSize() const = 0;
